controller.py内のinput()処理のバリデーションチェック関数をview.pyに移した。mode_select関数の定義
